Code/gRPC/requirements.txt
This file lists all required Python libraries.

grpcio
grpcio-tools
fastapi
uvicorn

Install Dependencies : pip install -r requirements.txt

